Refactor relationship compilation into separate module with improved modularity
